refactoring, ui updates and minor functionality fixes
