+++ src/js/Project/addProject.js
@@ -1,0 +1,6 @@
+export default class AddProject {
+  constructor(name) {
+    this._name = name;
+    this._todos = [];
+  }
+}
