--- src/js/addProject.js
@@ -1,8 +1,0 @@
-export default 
-
-class AddProject{
-    constructor(name){
-        this._name = name;
-        this._todos = [];
-    }
-}
