--- src/js/complete.js
@@ -1,3 +1,0 @@
-export default function completeTodo() {
-  console.log('complete');
-}
