--- src/js/priority.js
@@ -1,3 +1,0 @@
-export default function priority() {
-  console.log('priority');
-}
