+++ src/scss/style.scss
@@ -17,6 +17,14 @@ body {
   text-decoration: line-through;
 }
 
+a.bg-light:active {
+  background-color: #dae0e5 !important;
+}
+
+.w-70 {
+  width: 70% !important;
+}
+
 #sidebar-wrapper {
   min-height: 100vh;
   -webkit-transition: margin .25s ease-out;
